perf(web): lazy-load game routes and split Phaser/vendor chunks (#235)
Refactor each game's routes.tsx to lazy-load its route component body from a
sibling content.tsx via React.lazy + Suspense. registerRoutes() still runs at
startup (ADR-016 preserved) but only the lightweight route metadata lands in
the initial bundle.

Also add build.rollupOptions.output.manualChunks to split Phaser,
React/react-router, and @stitches/react into dedicated vendor chunks.

Before:  1 chunk (1,898.76 kB gzip 462.79 kB)
After:   phaser 1,478.41 kB (gzip 339.65 kB), vendor-react 161.76 kB
         (gzip 52.74 kB), vendor-stitches 17.09 kB, plus 17 per-game
         content chunks (8-27 kB each). App entry 11.25 kB (gzip 2.78 kB).

Closes #235
