+++ web/arcade/src/games/crunch3/content.tsx
@@ -1,0 +1,64 @@
+import { useState, useCallback } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
+import { GameCanvas } from '../../components/GameCanvas';
+import { GameHomeLayout } from '../../components/GameHomeLayout';
+import { StageMap, type StageInfo } from '../../components/StageMap';
+import { PlayLayout, isRN } from '../../components/PlayLayout';
+import { ClearScreen as Crunch3Clear } from './ClearScreen';
+import { HUD as Crunch3HUD } from './HUD';
+import { useGame as useCrunch3Game, type GameResult as Crunch3Result } from './useGame';
+
+const STAGES: StageInfo[] = Array.from({ length: 5 }, (_, i) => ({ id: i + 1, cleared: false }));
+
+export function Crunch3HomeRoute() {
+  const navigate = useNavigate();
+  return (
+    <GameHomeLayout title="Crunch 3" icon="🍕">
+      <StageMap
+        stages={STAGES}
+        currentStage={1}
+        onStageSelect={(stage) => navigate(`/games/crunch3/v1/stage/${stage}`)}
+      />
+      <p style={{ textAlign: 'center', padding: 8, fontSize: 12, color: '#9CA3AF' }}>Pixel food icons by Alex Kovacsart (CC BY 4.0)</p>
+    </GameHomeLayout>
+  );
+}
+
+export function Crunch3StageRoute() {
+  const { stageId } = useParams();
+  const navigate = useNavigate();
+  const parsed = parseInt(stageId || '1', 10);
+  const stage = Number.isNaN(parsed) ? 1 : parsed;
+  const [playKey, setPlayKey] = useState(0);
+  const [gameResult, setGameResult] = useState<Crunch3Result | null>(null);
+  const [screen, setScreen] = useState<'playing' | 'clear'>('playing');
+
+  const handleClear = useCallback((r: Crunch3Result) => {
+    if (!isRN) { setGameResult(r); setScreen('clear'); }
+  }, []);
+  const handleGameOver = useCallback((r: Crunch3Result) => {
+    if (!isRN) { setGameResult(r); setScreen('clear'); }
+  }, []);
+  const handleNext = useCallback(() => {
+    navigate(`/games/crunch3/v1/stage/${stage + 1}`, { replace: true });
+    setPlayKey((k) => k + 1); setScreen('playing');
+  }, [navigate, stage]);
+  const handleRetry = useCallback(() => { setPlayKey((k) => k + 1); setScreen('playing'); }, []);
+  const handleHome = useCallback(() => navigate('/games/crunch3/v1', { replace: true }), [navigate]);
+
+  if (screen === 'clear' && gameResult) {
+    return <Crunch3Clear result={gameResult} stage={stage} onNext={handleNext} onRetry={handleRetry} onHome={handleHome} />;
+  }
+
+  return <Crunch3Playing key={`${stage}-${playKey}`} stage={stage} onClear={handleClear} onGameOver={handleGameOver} />;
+}
+
+function Crunch3Playing({ stage, onClear, onGameOver }: { stage: number; onClear: (r: Crunch3Result) => void; onGameOver: (r: Crunch3Result) => void }) {
+  const { containerRef, score, combo, movesLeft, targetScore } = useCrunch3Game({ stage, onClear, onGameOver });
+  return (
+    <PlayLayout>
+      <Crunch3HUD stage={stage} score={score} targetScore={targetScore} movesLeft={movesLeft} combo={combo} />
+      <GameCanvas ref={containerRef} />
+    </PlayLayout>
+  );
+}
