+++ web/arcade/src/games/nexus-brawl/content.tsx
@@ -1,0 +1,15 @@
+export function NexusBrawlRoute() {
+  return (
+    <iframe
+      src="/nexus-brawl/index.html"
+      title="NEXUS BRAWL"
+      style={{
+        width: '100vw',
+        height: '100vh',
+        border: 0,
+        display: 'block',
+      }}
+      allow="autoplay; fullscreen"
+    />
+  );
+}
