+++ web/arcade/src/games/number10/routes.tsx
@@ -1,39 +1,28 @@
-import { useNavigate } from 'react-router-dom';
-import { GameCanvas } from '../../components/GameCanvas';
-import { GameHomeLayout } from '../../components/GameHomeLayout';
-import { PlayLayout } from '../../components/PlayLayout';
+import { lazy, Suspense } from 'react';
 import { registerRoutes } from '../../router';
-import { HUD as Number10HUD } from './HUD';
-import { useGame as useNumber10Game } from './useGame';
 
-function Number10HomeRoute() {
-  const navigate = useNavigate();
-  return (
-    <GameHomeLayout title="Make 10" icon="🔟">
-      <div style={{ flex: 1, display: 'flex', flexDirection: 'column', justifyContent: 'center', alignItems: 'center', gap: 16 }}>
-        <p style={{ fontSize: 16, color: '#6B7280' }}>Drag to select numbers that sum to 10!</p>
-        <button
-          onClick={() => navigate('/games/number10/v1/play')}
-          style={{ marginTop: 16, backgroundColor: '#2563EB', color: '#fff', border: 'none', padding: '18px 56px', borderRadius: 16, fontSize: 20, fontWeight: 700, cursor: 'pointer' }}
-        >
-          Play
-        </button>
-      </div>
-    </GameHomeLayout>
-  );
-}
-
-function Number10PlayRoute() {
-  const { containerRef, score, remaining } = useNumber10Game();
-  return (
-    <PlayLayout>
-      <Number10HUD score={score} remaining={remaining} />
-      <GameCanvas ref={containerRef} />
-    </PlayLayout>
-  );
-}
+const Number10HomeRoute = lazy(() =>
+  import('./content').then((m) => ({ default: m.Number10HomeRoute })),
+);
+const Number10PlayRoute = lazy(() =>
+  import('./content').then((m) => ({ default: m.Number10PlayRoute })),
+);
 
 registerRoutes('/games/number10/v1', [
-  { path: '', element: <Number10HomeRoute /> },
-  { path: 'play', element: <Number10PlayRoute /> },
+  {
+    path: '',
+    element: (
+      <Suspense fallback={null}>
+        <Number10HomeRoute />
+      </Suspense>
+    ),
+  },
+  {
+    path: 'play',
+    element: (
+      <Suspense fallback={null}>
+        <Number10PlayRoute />
+      </Suspense>
+    ),
+  },
 ]);
